--- client/src/App.js
@@ -1,16 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-
-import Home from "./components/pages/home";
-
-const App = (props) => {
-  return (
-    <Router>
-      <Switch>
-        <Route exact path="/" component={Home} />
-      </Switch>
-    </Router>
-  );
-};
-
-export default App;
